split generate to more readable chunks
